+++ src/objects/impls/creep.rs
@@ -11,7 +11,7 @@ use crate::{
     CostMatrix, MoveToOptions, RoomName, RoomPosition,
 };
 
-#[cfg(feature = "thorium")]
+#[cfg(feature = "seasonal-season-5")]
 use crate::objects::Reactor;
 
 #[wasm_bindgen]
@@ -83,7 +83,7 @@ extern "C" {
     #[wasm_bindgen(final, method, js_name = claimController)]
     fn claim_controller_internal(this: &Creep, target: &StructureController) -> i8;
 
-    #[cfg(feature = "thorium")]
+    #[cfg(feature = "seasonal-season-5")]
     #[wasm_bindgen(final, method, js_name = claimReactor)]
     fn claim_reactor_internal(this: &Creep, target: &Reactor) -> i8;
 
@@ -306,7 +306,7 @@ impl Creep {
     /// parts.
     ///
     /// [Screeps documentation](https://docs-season.screeps.com/api/#Creep.claimReactor)
-    #[cfg(feature = "thorium")]
+    #[cfg(feature = "seasonal-season-5")]
     pub fn claim_reactor(&self, target: &Reactor) -> Result<(), ErrorCode> {
         ErrorCode::result_from_i8(self.claim_reactor_internal(target))
     }
@@ -457,8 +457,11 @@ impl Creep {
     /// parts.
     ///
     /// [Screeps documentation](https://docs.screeps.com/api/#Creep.repair)
-    pub fn repair(&self, target: &RoomObject) -> Result<(), ErrorCode> {
-        ErrorCode::result_from_i8(self.repair_internal(target))
+    pub fn repair<T>(&self, target: &T) -> Result<(), ErrorCode>
+    where
+        T: ?Sized + Repairable,
+    {
+        ErrorCode::result_from_i8(self.repair_internal(target.as_ref()))
     }
 
     /// Reserve an unowned [`StructureController`] in melee range using a
@@ -523,8 +526,12 @@ impl HasHits for Creep {
     }
 }
 
-impl MaybeHasNativeId for Creep {
-    fn try_native_id(&self) -> Option<JsString> {
+impl MaybeHasId for Creep {
+    /// The Object ID of the [`Creep`], or `None` if it began spawning this
+    /// tick.
+    ///
+    /// [Screeps documentation](https://docs.screeps.com/api/#Creep.id)
+    fn try_js_raw_id(&self) -> Option<JsString> {
         self.id_internal()
     }
 }
@@ -534,6 +541,10 @@ impl HasStore for Creep {
         self.store()
     }
 }
+
+impl Attackable for Creep {}
+impl Healable for Creep {}
+impl Transferable for Creep {}
 
 impl SharedCreepProperties for Creep {
     fn memory(&self) -> JsValue {